fix: upgrade promisify-child-process
BREAKING CHANGE: If you want stdout or stderr in the result of a promake.spawned child process, you
must now pass the maxBuffer or encoding option. (maxBuffer defaults to 200 * 1024 if only encoding
is given).  In previous versions all output was always captured if piped to the parent, potentially
exhausting available memory and crashing the parent process.